pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "elevenlabs-toolkit"
version = "0.1.0"
description = "ElevenLabs transcription and transcript transformation toolkit"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "elevenlabs",
  "tqdm",
  "python-dotenv",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
